Module serde_closure::structs

Structs representing a serializable closure, created by the FnOnce, FnMut and Fn macros. They implement std::ops::FnOnce, std::ops::FnMut and std::ops::Fn respectively, as well as Debug, Serialize and Deserialize, and various convenience traits.

See the readme for examples.

Structs

Fn

A struct representing a serializable closure, created by the Fn macro. Implements std::ops::Fn, Debug, Serialize and Deserialize, and various convenience traits.

FnMut

A struct representing a serializable closure, created by the FnMut macro. Implements std::ops::FnMut, Debug, Serialize and Deserialize, and various convenience traits.

FnOnce

A struct representing a serializable closure, created by the FnOnce macro. Implements std::ops::FnOnce, Debug, Serialize and Deserialize, and various convenience traits.