Module serde_closure::traits

Supertraits of std::ops::FnOnce, std::ops::FnMut and std::ops::Fn that are usable on stable Rust. They are implemented by closures created by the FnOnce, FnMut and Fn macros.

See the readme for examples.

Traits

Fn

Supertrait of std::ops::Fn that is usable on stable Rust. It is implemented by closures created by the Fn macro.

FnMut

Supertrait of std::ops::FnMut that is usable on stable Rust. It is implemented by closures created by the FnMut macro.

FnOnce

Supertrait of std::ops::FnOnce that is usable on stable Rust. It is implemented by closures created by the FnOnce macro.