Trait serde_closure::traits::FnOnce

pub trait FnOnce<Args> {
    type Output;

    fn call_once(self, args: Args) -> Self::Output;
}

Supertrait of std::ops::FnOnce that is usable on stable Rust. It is implemented by closures created by the FnOnce macro.

See the readme for examples.

Required Associated Types

type Output

The returned type after the call operator is used.

Required Methods

fn call_once(self, args: Args) -> Self::Output

Performs the call operation.